feat(step2) : lotto 여러개 구입
