+++ src/main/java/lotto/domain/Lottos.java
@@ -3,9 +3,9 @@ package lotto.domain;
 import java.util.List;
 
 public class Lottos {
-    private List<Lotto> lottos;
+    private final List<Lotto> lottos;
 
-    public Lottos(List<Lotto> lottos) {
+    public Lottos(final List<Lotto> lottos) {
         this.lottos = lottos;
     }
