--- src/main/java/lotto/domain/PlayLotto.java
@@ -1,11 +1,0 @@
-package lotto.domain;
-
-import java.util.List;
-
-public class PlayLotto {
-    private Lottos lottos;
-    private int count;
-
-
-
-}
